+++ dqops/src/main/java/com/dqops/sensors/CustomSensorParametersSpec.java
@@ -22,6 +22,7 @@ import com.dqops.metadata.id.ChildHierarchyNodeFieldMap;
 import com.dqops.metadata.id.ChildHierarchyNodeFieldMapImpl;
 import com.dqops.utils.schema.JsonAdditionalProperties;
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -150,5 +151,17 @@ public class CustomSensorParametersSpec extends AbstractSensorParametersSpec imp
     @Override
     public String getSensorDefinitionName() {
         throw new UnsupportedOperationException("A custom sensor parameters does not known its sensor name, it is defined on the custom check definition.");
+    }
+
+    /**
+     * Checks if the object is a default value, so it would be rendered as an empty node. We want to skip it and not render it to YAML.
+     * The implementation of this interface method should check all object's fields to find if at least one of them has a non-default value or is not null, so it should be rendered.
+     *
+     * @return true when the object has the default values only and should not be rendered to YAML, false when it should be rendered.
+     */
+    @Override
+    @JsonIgnore
+    public boolean isDefault() {
+        return super.getAdditionalProperties() == null || super.getAdditionalProperties().isEmpty();
     }
 }