+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.11.2
+# 1.11.3
 
-* Data synchronization preserves remote changes when local files were also modified
-* Support repeated (non-array) values in BigQuery after some schema changes in the INFORMATION_SCHEMA.COLUMNS management view.
+* Fixes to use parameters for custom checks.
+* Shorten the time in the connection pool to 10 sec